Maybe make tests less racy by deleting key
This may be relevant if a test failed on the previous iteration.

This commit also replaces all `Result::is_ok` and `Result::is_err` calls
in tests with `.unwrap()` and `.unwrap_err()` respectively. This has the
advantage of also printing out what (unexpected) value we *did* get if
the assertion fails.

Unfortunately, some of the tests are still racy, but I'm not entirely
sure why...